Added some tests in reset_app_prefix script to avoid spurious error messages.

+++ app/reset_app_prefix.sh
@@ -8,14 +8,15 @@ STACKNAME=$(basename $PARENTDIR)
 # echo $BASEDIRECTORY
 # echo $STACKNAME
 
-source $PARENTDIR/.env
-echo $DOMAIN
-echo $PREFIX
-echo $ENVIRONNEMENT
+if [ -f .env ]; then
+    SED_CMD='s/APP_PREFIX=.*/APP_PREFIX=""/g'
+    sed -i $SED_CMD .env
+fi
 
-SED_CMD='s/APP_PREFIX=.*/APP_PREFIX=""/g'
-sed -i $SED_CMD .env
+if [ -f $PARENTDIR/.env ]; then
+    source $PARENTDIR/.env
+    cd public
+    rm -f $PREFIX
+fi
 
 
-cd public
-rm -f $PREFIX
